Updated the way info sidebar to be more efficient with re-population. Split the sidebar's tag table  into two groups for relevant and irrelevant tags. Made small styling changes
