+++ src/scripts/popup/index.ts
@@ -2,13 +2,8 @@ import { FontAwesomeIcon } from "../components/icon.js";
 import { ElementBuilder } from "../elements.js";
 import { State } from "../state/index.js";
 import { createCustomElement, getElement } from "../supplement/elements.js";
-import { OsmValue } from "../types/osm.js";
 
-export const POPUP = getElement("popup", HTMLDialogElement);
-export const WAY_INFO = getElement("way-info", HTMLDivElement);
-export const WAY_INFO_ID = getElement("wayid", HTMLHeadingElement);
-export const WAY_INFO_LINK = getElement("wayid-link", HTMLAnchorElement);
-export const WAY_INFO_TAGS = getElement("tags", HTMLTableElement);
+export const POPUP_DIALOG = getElement("popup", HTMLDialogElement);
 
 export abstract class Popup {
 	protected abstract title: string;
@@ -24,10 +19,10 @@ export abstract class Popup {
 
 	display() {
 		// purge all existing children
-		while (POPUP.lastChild !== null) POPUP.lastChild.remove();
+		while (POPUP_DIALOG.lastChild !== null) POPUP_DIALOG.lastChild.remove();
 
 		// add window popup decorations
-		const heading = new ElementBuilder("h3").text(this.title).build();
+		const heading = new ElementBuilder("h2").text(this.title).build();
 		const closeIcon = createCustomElement(FontAwesomeIcon).setIcon("xmark");
 		const closeButton = new ElementBuilder("button")
 			.id("popup-close")
@@ -42,89 +37,15 @@ export abstract class Popup {
 		const main = new ElementBuilder("main").children(...this.children).build();
 
 		// show popup
-		POPUP.append(header, main);
-		POPUP.showModal();
-		POPUP.scrollTop = 0;
+		POPUP_DIALOG.append(header, main);
+		POPUP_DIALOG.showModal();
+		POPUP_DIALOG.scrollTop = 0;
 	}
 
 	static close() {
-		if (!POPUP.open) return;
-		POPUP.close();
+		if (!POPUP_DIALOG.open) return;
+		POPUP_DIALOG.close();
 	}
-}
-
-export function displaySidebar(wayId: number) {
-	// get data for way
-	const wayData = State.data.get()?.get(wayId);
-	if (wayData === undefined) return;
-
-	// update heading url and text
-	WAY_INFO_LINK.href = `https://www.openstreetmap.org/way/${wayId}`;
-	WAY_INFO_LINK.textContent = wayId.toString();
-
-	// purge all tag entries from any previous ways
-	while (WAY_INFO_TAGS.lastChild) WAY_INFO_TAGS.removeChild(WAY_INFO_TAGS.lastChild);
-
-	// create heading row
-	const tagHeading = new ElementBuilder("th").text("Tag").build();
-	const valueHeading = new ElementBuilder("th").text("Value").build();
-	const row = new ElementBuilder("tr").children(tagHeading, valueHeading).build();
-	WAY_INFO_TAGS.append(row);
-
-	// compile all original tags
-	const originalTags = wayData.originalWay.tags ?? {};
-	const originalTagMap = new Map<string, string>();
-	Object.entries(originalTags).forEach(([tag, value]) => originalTagMap.set(tag, value));
-
-	// compile all inferred tags
-	const inferredTags = wayData.tags;
-	const inferredTagMap = new Map<string, string>();
-	Object.entries(inferredTags).forEach(([tag, value]) => {
-		// format tag
-		const words = Array.from(tag);
-		let formattedTag = "";
-		for (let i = 0; i < words.length; i++) {
-			const letter = words[i];
-			if (letter.toUpperCase() !== letter) formattedTag += letter;
-			else formattedTag += `:${letter.toLowerCase()}`;
-		}
-
-		// format value
-		let valueString = "<unknown>";
-		if (value instanceof OsmValue) valueString = value.toString();
-
-		inferredTagMap.set(formattedTag, valueString);
-	});
-
-	// get all tags from both maps
-	const allOriginalTags = Array.from(originalTagMap.keys());
-	const allInferredTags = Array.from(inferredTagMap.keys());
-	const allTags = new Set([...allOriginalTags, ...allInferredTags].sort());
-
-	// create each content row
-	for (const tag of allTags) {
-		const originalValue = originalTagMap.get(tag);
-		const inferredValue = inferredTagMap.get(tag);
-
-		// generate display string
-		let displayString = originalValue;
-		if (displayString === undefined && inferredValue !== undefined) {
-			if (inferredValue === "") displayString = "<no value>";
-			else displayString = `${inferredValue} (inferred)`;
-		}
-		displayString ??= "<unknown>";
-
-		// build row
-		const tagCell = new ElementBuilder("td").text(tag.toString()).build();
-		const valueCell = new ElementBuilder("td").text(displayString).build();
-		const tagRow = new ElementBuilder("tr").children(tagCell, valueCell).build();
-
-		WAY_INFO_TAGS.append(tagRow);
-	}
-
-	WAY_INFO.removeAttribute("hidden");
-	State.selectedWay.set(wayId);
-	WAY_INFO.scrollTop = 0;
 }
 
 export function openID() {
