+++ src/styles/base.scss
@@ -57,7 +57,7 @@
 	}
 }
 
-// # fonts
+// # text & fonts
 h1,
 h2,
 h3,
@@ -80,22 +80,24 @@ h3,
 	margin-block: 0.25em 0;
 
 	font-family: "TekTur", sans-serif;
-	font-weight: 600;
 }
 
 h1,
 .h1 {
 	font-size: 2rem;
+	font-weight: 800;
 }
 
 h2,
 .h2 {
 	font-size: 1.75rem;
+	font-weight: 700;
 }
 
 h3,
 .h3 {
-	font-size: 1.5rem;
+	font-size: 1.375rem;
+	font-weight: 600;
 }
 
 p,
@@ -112,6 +114,10 @@ table {
 code {
 	font-family: "Fira Code", monospace;
 	font-weight: 300;
+}
+
+ul {
+	padding-inline-start: 2ch;
 }
 
 // # buttons & links
